Extracted BepStreamObserver class
